Updating the Indexter cognitive filter to sort steps before terminating
The plan recognition pipeline is sensitive to the input order of the observations, which I had not accounted for in the original design of the algorithm.  Now, before terminating, the Indexter filter sorts the salient steps by step id (i.e. sorts the salient steps into the relative order they appear in the chronology).